automatic_monitoring_skyeye/requirements.txt
baidu-aip==2.2.18.0
certifi==2019.11.28
chardet==3.0.4
configparser==4.0.2
idna==2.8
lxml==4.4.2
Pillow==7.0.0
pytesseract==0.3.2
requests==2.22.0
selenium==3.141.0
tesseract==0.1.3
urllib3==1.25.8
